% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_paste.R
\name{\%paste\%}
\alias{\%paste\%}
\title{Paste operator}
\usage{
x \%paste\% y
}
\arguments{
\item{x}{First string}

\item{y}{Second string}
}
\value{
x followed by y
}
\description{
An infix operator to concatenate strings
}
