% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkmisc-package.R
\docType{package}
\name{rkmisc-package}
\alias{rkmisc}
\alias{rkmisc-package}
\title{rkmisc: Helps RK Use R}
\description{
A collection of R functions Robert Karlsson thought would be useful
    at some point.
}
\details{
These are details from rkmisc-package.R - anything else is from DESCRIPTION
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robkar/rkmisc}
  \item Report bugs at \url{https://github.com/robkar/rkmisc/issues}
}

}
\author{
\strong{Maintainer}: Robert Karlsson \email{robkar@gmail.com} (0000-0002-8949-2587)

}
\keyword{internal}
